Moved directory and fixed minitest plugins.
