+++ lib/util/client.rb
