+++ lib/util/client/client.rb
