+++ lib/util/client/grpc.rb
