+++ lib/util/client/grpc/client.rb
