+++ lib/util/client/utils.rb
