+++ lib/util/constants.rb
