+++ lib/util/environment.rb
