+++ lib/util/exceptions.rb
