+++ lib/util/logger.rb
