+++ lib/util/node.rb
