+++ lib/util/node_util.rb
