+++ tests/minitest/test_netconf_yang.rb
@@ -19,7 +19,7 @@
 
 require 'rexml/document'
 require_relative 'ciscotest'
-require_relative '../../lib/puppet/node_utils/client/netconf/netconf'
+require_relative '../../lib/util/client/netconf/netconf'
 
 # TestNetconf- Minitest for Netconf class
 class TestNetconf < CiscoTestCase
